server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Search & scraping
tavily-python==0.3.3
requests==2.31.0
beautifulsoup4==4.12.2
trafilatura==1.6.4

# ML/AI (if you need them)
google-generativeai==0.3.2